renamed files to not be BEM and be .jsx on comps
